refactor(qnaBoard): QnaBoardService 인터페이스 삭제했습니다.
